master

+++ backend/routes/userRouter.js
@@ -7,42 +7,52 @@ const auth = require("../middleware/auth");
 // Register route
 router.post("/register", async (req, res) => {
   try {
-    const { id, password, passwordCheck } = req.body;
 
-    // validate
-    // status code 400 means bad request
-    // status code 500 means internal server error
+    const token = req.header("x-auth-token");
+    if (!token) return res.json(false);
 
-    if (!id || !password || !passwordCheck) {
-      return res.status(400).json({ msg: "Not all fields have been entered" });
-    }
+    const verified = jwt.verify(token, process.env.JWT_SECRET);
+    if (!verified) return res.json(false);
 
-    // Checking to ensure password length is at least 5 characters
-    if (password.length < process.env.MIN_PASSWORD_LENGTH) {
-      return res
-        .status(400)
-        .json({ msg: `The password needs to be at least ${process.env.MIN_PASSWORD_LENGTH} characters long` });
-    }
+    const user = await User.findById(verified.id);
+    if (!user) return res.json(false);
 
-    // Checking the password entered vs the password checker
-    if (password !== passwordCheck) {
-      return res
-        .status(400)
-        .json({ msg: "Passwords do not match. Please try again" });
-    }
+    if (!user.isAdmin) {res.status(401).json({msg: "Admin privileges required"}); return;}
+      const { id, password, passwordCheck } = req.body;
 
-    // using Bcrypt to hash passwords for security
-    const salt = await bcrypt.genSalt();
-    const passwordHash = await bcrypt.hash(password, salt);
+      // validate
+      // status code 400 means bad request
+      // status code 500 means internal server error
 
-    // creating out new user notice password value is passwordHash not password
-    const newUser = new User({
-      id: id,
-      password: passwordHash,
-    });
-    const savedUser = await newUser.save();
-    res.json(savedUser);
+      if (!id || !password || !passwordCheck) {
+        return res.status(400).json({ msg: "Not all fields have been entered" });
+      }
 
+      // Checking to ensure password length is at least 5 characters
+      if (password.length < process.env.MIN_PASSWORD_LENGTH) {
+        return res
+          .status(400)
+          .json({ msg: `The password needs to be at least ${process.env.MIN_PASSWORD_LENGTH} characters long` });
+      }
+
+      // Checking the password entered vs the password checker
+      if (password !== passwordCheck) {
+        return res
+          .status(400)
+          .json({ msg: "Passwords do not match. Please try again" });
+      }
+
+      // using Bcrypt to hash passwords for security
+      const salt = await bcrypt.genSalt();
+      const passwordHash = await bcrypt.hash(password, salt);
+
+      // creating out new user notice password value is passwordHash not password
+      const newUser = new User({
+        id: id,
+        password: passwordHash,
+      });
+      const savedUser = await newUser.save();
+      res.json(savedUser);
     // Catching any errors that come through
   } catch (error) {
     res.status(500).json({ err: error.message });
@@ -92,6 +102,16 @@ router.post("/login", async (req, res) => {
 // delete user account route
 router.delete("/delete", auth, async (req, res) => {
   try {
+    const token = req.header("x-auth-token");
+    if (!token) return res.json(false);
+
+    const verified = jwt.verify(token, process.env.JWT_SECRET);
+    if (!verified) return res.json(false);
+
+    const user = await User.findById(verified.id);
+    if (!user) return res.json(false);
+
+    if(!user.isAdmin) { return res.status(401).json({ msg: "You do not have permission to do this" }); }
     const deletedUser = await User.findByIdAndDelete(req.user);
     res.json(deletedUser);
   } catch (error) {
@@ -116,26 +136,26 @@ router.post("/tokenIsValid", async (req, res) => {
     res.status(500).json({ err: error.message });
   }
 });
- 
-router.pos("/isAdmin", async(req, res) => {
-    try {
-        const token = req.header("x-auth-token");
-        if (!token) return res.json(false);
 
-        const verified = jwt.verify(token, process.env.JWT_SECRET);
-        if (!verified) return res.json(false);
+router.post("/isAdmin", async (req, res) => {
+  try {
+    const token = req.header("x-auth-token");
+    if (!token) return res.json(false);
 
-        const user = await User.findById(verified.id);
-        if (!user) return res.json(false);
+    const verified = jwt.verify(token, process.env.JWT_SECRET);
+    if (!verified) return res.json(false);
 
-        return res.json(user.isAdmin);
-    } catch (error) {
-        res.status(500).json({ err: error.message });
-    }
+    const user = await User.findById(verified.id);
+    if (!user) return res.json(false);
+
+    return res.json(user.isAdmin);
+  } catch (error) {
+    res.status(500).json({ err: error.message });
+  }
 });
 
 // This route is grabbing one user
-router.get("/", auth, async (req,res) => {
+router.get("/", auth, async (req, res) => {
   const user = await User.findById(req.user)
   res.json({
     firstName: user.firstName,
